+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
   - Support id parameter for pageUrl. (CMS-4950)
   - Better componentUrl handling. Support "this" component and component to other pages. (CMS-4995)
   - Improvement to imageUrl function. Allow id and path to image. (CMS-5002)
+  - Improvement to attachmentUrl function. Resolve label to name and allow id/path to content. (CMS-4951)
 
 Refactoring: